special mappings for <CR> and <Tab>
If autoclose inserted the corresponding character and you are between the
two characters ( ex: between a pair of `"` ), pressing Enter will open a line
below the current one and will put the cursor at the beginning. Pressing Tab
will move the cursor a space after the closing character.

Thus, instead of havint to press Escape ( or, god forbid, the arrow keys )
and moving past the closing character, you can do that by pressing a single key.